// src/types/style.ts
export interface Style {
  width?: string;
  height?: string;
  flexDirection?: string;
  display?: string;
  gap?: string;
  alignItems?: string;
  justifyContent?: string;
  padding?: string;
  margin?: string;
  backgroundColor?: string;
  overflow?: string;
  overflowY?: string;
  overflowX?: string;
  tabletWidth?: string;
  mobileWidth?:string;
}